Move helper code to concern
